16th commit: SHA256TokenHash functionality add.
A SHA256TokenHash.java class which implements an interface called HashStrategy.java introduced to the project. it is placed in hash sub-package under util package. SHA256TokenHash is the primary hash strategy for now, that's the reason for annotating it with @Primary annotation. To be more cautious as well as following Dependency Inversion Principle a new layer of abstraction meaning HashStrategy.java interface is added here.

I set SHA256TokenHash.java to package private in order to access hashing functionality through HashService.java class therefor, HashService.java service class added to the project.

perhaps I should mention that this hashing functionality will be used during login phase while saving refresh token in the database and a few other places which I'm not aware of it yet, refresh token will be added to the project later.