19th commit: saveRefreshToken method implementation added.
Saving RefreshToken functionality introduced here in order to save the state of RefreshToken as well as to revoke and validate it in the future scenarios. This method naturally takes a number of arguments such as username,hashedToken,deviceId etc.

Initially I check the username against DB for its existence and if it did not exist it throws an exception with an informative message. In order to enforce one active refresh token per device I check the user+deviceId against DB and delete all existingRefreshTokens. Then I create A RefreshToken Object through refreshTokenFactory.java class and finally save the RefreshToken into the DB.

I annotated the saveRefreshToken method to be @Transactional for now to make DB operations atomic it may be subject to change.

I edited the method name of refreshTokenFactory to create which was previously titled with save, as I thought it was causing ambiguity before calling refreshTokenRepository.save() method.