22nd commit: another level of RefreshToken Validity against DB added to JwtTokenUtil class.
Another level of validation for RefreshToken against DB added inside JwtTokenUtil.java class which is named validateRefreshTokenFingerPrintsExpiryAgainstDB method, it takes four arguments including one for hashedRefreshToken and the remaining are related to device fingerprints such as deviceId,userAgent and ipAddress.

The validateRefreshTokenFingerPrintsExpiryAgainstDB() method is called in validateRefreshToken(), using isTokenValid() method of refreshTokenService it checks the validity of hashedRefreshToken or else it throws an InvalidTokenException with informative message.