29th commit: Central Authentication controller added
1_AuthController is a restController (@ResponseBody+@Controller or @RestController) which handles 4 endpoints for now ,it is not oauth2 compatible yet as I need to change a few endpoints' names, for instance "/register" should be changed to "/token".

This controller injects an AuthenticationService to call its methods, I tried to keep the AuthenticationController.java class as clean as possible as controllers are only responsible to call business logic of the services and also perform some validation logic which will be done later.

2_I added AuthenticationService.java interface which is a public interface and its implementation class AuthenticationServiceImpl.java class which is protected (package private). This class needs more work to do that what I've done now such as bruteforce protection which will be added later.

For the register() method what I have done is , it takes an AuthenticationRequest as its argument, checks the user in DB through UserService (and not userRepository) ,if there was no duplicate it sets a default "ROLE_USER" to the user.Finally, it will return an AuthenticationResponse to the user after successful registration.

###Based on my understandings, AuthenticationServers (in case of registering users) are only responsible to set default USER_ROLE and create regular users and not admin users, therefor I will set the admin user with ADMIN_ROLE in the initializer class later.

###There are a billion ways to set roles dynamically, I try to keep it simple for now.

3_Also I added a RegistrationException.java exception class inside exception packages which is handled in CentralExceptionHandler.java handler class.