30th commit: AuthenticationServiceImpl class methods almost implemented.
1_login method implemented
Login method in AuthenticationServiceImpl class takes two arguments including authenticationRequest and httpServletRequest, I injected an authenticationManager in this class to create an unauthenticated Authentication object with an implementation of it which is UsernamePasswordAuthenticationToken, the process is in such way that the AuthenticationManagers checks the CustomAuthenticationProvider.java class to validate credentials against the stored user in DB and perform authentication logic. Next step I gathered the user fingerPrints including ip,deviceId and its agent and then did generate both access and refreshTokens , hashed the refreshToken and saved it in the DB for revocation and validation. finally returned a AuthenticationResponse to the user with an informative message.

I must change DB strategy to redis cash for faster speed and .... in the near future.

2_logout method implemented.
Logout method takes two arguments as the login method does, here I gathered the user fingerPrints, validated the refreshToken using jwtTokenUtil.java class, hashed the token and revoked it in the DB, here revoked means deletion and I deleted the whole refresh token that belongs to the user.Finally, I returned an AuthenticationResponse with an informative message to the user.

For more future related tasks like bruteforce protection I also updated register method signature in 3 places including AuthenticationController.java , AuthenticationService.java and AuthenticationServiceImpl.java classes (added HttpServletRequest argument to register() method signature).

3_Lastly, I deleted HttpContextIpProvider as it bound the code to the RequestContextHolder.No need of this class anymore and changed ClientIpService title to IpAddressResolverService.