38th commit: Email Mobile password validation strategies added
A number of validation strategy classes and interfaces defined under strategy sub-package for email,mobile and password as follows:

EmailBlank
EmailLength
which implement EmailValidationStrategy.java interface.

MobileBlank
MobileLength
which implement MobileValidationStrategy.java interface.

PasswordBlank
PasswordLength
PasswordComplexity
which implement PasswordValidationStrategy.java interface.

A method called isValid(String input) has exactly same signature between all the implementation classes but the type of interface they implement is totally different. isValid() method is responsible to validate the input based on a specific criteria. For instance in EmailBlank implementation class it checks whether if the entered email is not null( blank ) or it is not null, if the email is valid it will return an Optional.empty() and if email is not valid meaning it is blank or null it will throw a ValidationError.java.

ValidationError.java is custom exception class defined under exception->error package.