40th commit: custom constraintValidator classes for username password added.
1_Two custom constraintValidator classes added to support validPassword and ValidUsername annotations, they intended two define how the validation process should be done for username and password.

These to custom constraint validators are placed under validator sub-package of util->Validation.

The first one PasswordConstraintValidator.java overrides the isValid() method of ConstraintValidator and is dependent on a List<PasswordValidationStrategy>. It makes a stream of passwordValidationStrategies and map the result of each strategy.isValid(password) to a strategy (or ValidationError) and filters if any ValidationError is present, if a validationError is present it takes the first one and returns it or else it return an Optional.empty().

When it returns an optional.empty() it means the entered password is in compliance with the validation rules the method should return true. If the stream contains ValidationError and the validationError is present in the if block we disable disableDefaultConstraintViolation and build a new buildConstraintViolationWithTemplate with the error message we get from validationError and the method return false.

To display this Validation error we should handle a MethodArgumentNotValidException in the CentralExceptionHandler which is implemented.

The UsernameConstraintValidator is a bit different since it is dependent on a UsernameValidationOrchestrator, it makes a stream of usernameValidationOrchestrators, filters the stream to find out the username is supported with which orchestrator (either EmailValidationOrchestrator or MobileValidationOrchestrator), it finds the first implementation that supports the username and calls isValid() method of the implementation class to validate the username. The rest of the process is the same as the PasswordConstraintValidator regarding custom error messages.