41st commit: Null handling of request fingerprints with default values added.
1_In order to prevent Null values (user fingerprints) from being entered into the system by mistake A record called AuthenticationRequestFingerprint (immutable values) added under dto sub-package of authentication package that holds fingerprints' values.

2_A method called extractAuthenticationRequestFingerprint which takes HttpServletRequest as its argument added to AuthenticationServiceImpl class. This method is intended to extract ipAddress,deviceId and lastly userAgent from HttpServletRequest and reused in multiple methods.

This method is a centralized extraction logic which safely handles potential null since httpServletRequest.getHeader("X-Device-Id") may return either the actual value of X-Device-Id or a null value.

The filter checks if the header value is none blank and passes and if it is blank is create an Optional.empty(), .orElse("UNKNOWN_DEVICE_ID") will return the value inside Optional if it is present, or it falls back to "UNKNOWN_DEVICE_ID if header was null or blank.

3_The process is exactly the same for userAgent.

4_The return type Logout method in AuthenticationController,AuthenticationService and AuthenticationServiceImpl classes and interfaces updates with a new return type LogoutResponse which is a record.

5_AuthenticationResponse,RegistrationResponse,AuthenticationRequest all changed to record instead of class.

6_Custom annotations related to validation updated with @Target({ElementType.PARAMETER, ElementType.FIELD})