43rd commit: Custom RateLimit for controller methods added.
1_ Declarative Rate limiting was introduced into the project to prevent abuse as well as provide far usage of the controller methods. All the classes and interfaces related to rate limiting functionality are placed under

authentication -> security -> ratelimit package.

A) This functionality starts off with an Annotation called RateLimit which is applicable to methods, it takes four attributes including key,maxRequests,timeFrameInMinutes an enum type called identifiers that can be both IP and User or just one.

B) RateLimit is dependent on resolving identifiers (IdentifierResolver) which has two implementations, IpIdentifierResolver and UserIdentifierResolver.

C) RateLimiterKeyBuilder is the central (component or service) which is responsible to create redis key based on combination of IP and User or just one of these two attributes.

D_ The final part is the aspect which is responsible for managing Redis cache functionality based on @RateLimit and key that is built through keyBuilder.java.

2_ Redis cache config added to properties files.