44th commit: service layer validation (CredentialValidationService) added.
1_Service layer validation intended to prevent faulty data into the system as the second level of validation, in the first level meaning declarative validation layer I prevent the users' requests to go further down the chain of processes using @ValidUsername and @ValidPassword, therefor the faulty invalid credentials will be stopped in the Controller layer.

The CredentialValidationService.java interface and its implementation class added to a sub-package called validator under the authentication package. This service is dependent on UsernameValidator and PasswordValidator the two validation components that are placed under util->validation-> validator package. In the CredentialValidationServiceImpl if any error happens in any of these validator components that error or errors will be added to a List of ValidationError type and then using a new error class called CredentialValidationException, these errors will be thrown to the user and then will be handled using the CentralExceptionHandler advice.

The CredentialValidationException is a custom exception and is placed under error sub-package of exception package.

Lastly, the validate method of credentialValidationService will be called inside register method and takes an authenticationRequest as its argument.

The main validation processes will be happened in the PasswordValidator and UsernameValidator components.For instance, in the validate method of the PasswordValidator, it takes the password entered by the user, then it creates a stream of the PasswordValidationStrategy and each strategy is an object which can validate password based on a certain criteria that will return either a ValidationError or an Optional.Empty(), next step using a map it applies the password to isValid() method of each strategy, then using filter it keeps the Optional of non-empty (errors) if any existed and finds the first non-empty ValidationError otherwise it will return an Optional.empty() which means no error happened.

The UsernameValidator is almost similar to the PasswordValidator but with a minor difference, the validate() method of the UsernameValidator component creates a stream of type UsernameValidationOrchestrator (email or mobile validationOrchestrator), it applies a filter to the support method of email or mobile Orchestrators , then finds the first orchestrator which supports handling email or mobile username types for validation, and using map it applies username to isValid method of the strategies that can validate that username type.

For instance, the email orchestrator is the one that supports the username entered by the user to validate, then using map the isValid() method of the EmailValidationOrchestrator will be called, and the EmailValidationOrchestrator has a list of emailValidationStrategies to validate email based on some criteria.