45th commit: brute-force protection added to CustomAuthenticationProvider for login process.
brute-force protection service intended to mitigate a cyberattack type in which the attackers try every combination of characters to guess username and passwords. This security measure starts off by defining an interface called BruteForce and its implementation class called RedisBruteForce as well as a record type called BruteForcePolicy.

I also already added two property classes called LoginBruteForceProperties and RegisterBruteForceProperties, explanation of all these classes already mentioned on previous commits.

1_When a user requested to log in through the controller login method, this method calls the login method in the AuthenticationServiceImpl.java class, then in the execution flow of the login process, an unauthenticated Authentication object with the user's credentials will be created and then the authenticationManager calls authenticate method with unauthenticated Authentication object.

Next step providerManager iterates through registered authenticationProviders and since my CustomAuthenticationProvider supports UsernamePasswordAuthenticationToken , it will respond true to the providerManager and the providerManager delegates authentication process to the CustomAuthenticationProvider.

note: sine I don't have direct access to HttpServletRequest in the CustomAuthenticationProvider, I added a component class called HttpContextIpAddressResolver which is dependent on IpAddressResolverService to detect the ip address of client using RequestContextHolder.

In the authenticate() method of the CustomAuthenticationProvider, I first checked whether if the user already blocked or not, if it is not then in the try catch block it check the user in the DB as:
 userDetailsService.loadUserByUsername(username);

 A)If the user does not exist in the DB I catch the UsernameNotFoundException which is thrown by UserService and add failure, if the user existed I check whether if the entered password by the user matches with the password in the DB, if not I add a failure to redis.

 B)If everything goes smooth I remove any lock by calling loginBruteForceService.onSuccess(ip,username);

 Important:
 note that in all exceptions I threw  BadCredentialsException("error: Invalid Credentials."); for security reason.

 (issue: the authenticate() method in CustomAuthenticationProvider called twice , it will be solved hopefully! )