46th commit: CustomAuthenticationProvider dual invoking solved
In my previous updates I declared the CustomAuthenticationProvider in the SecurityConfig as an object attribute and used to register it with AuthenticationManagerBuilder.java class and registering the CustomAuthenticationProvider in SecurityConfig.java class as well as setting the CustomAuthenticationProvider as a @Component that implements AuthenticationProvider was the issue which caused Spring security to assume that there are two available AuthenticationProvider in the project.

note: I also remove .oauth2ResourceServer(oauth2 -> oauth2.jwt(jwt -> jwt.decoder(jwtDecoder))) from the SecurityConfig.java config file since the role of this authentication server is to issue tokens, login , logout and register new users.