53rd commit: Allow logout and rotate endpoints with refresh token
logout and rotate endpoints do not need to be secured with access token (no need to be authenticated with access token)

In order to access /rotate and /logout endpoints I let these two endpoints to be permitted to all , sending access token in the request header was unnecessary and definitely causes deadlock, since access token is short-lived token (15 minutes or so) , therefor I set these two endpoint to permitted in the SecurityConfig.java class.

RefreshToken acts as the proof of identity which is already bounded to request ipAddress,deviceId and userAgent.

1_moved AuthenticationRequestFingerprint to security sub-package under authentication package inside fingerprint package

2_AuthenticationRequestFingerprintExtractor defined with a single method called extractFrom()

3_A new implementation class defined that is called AuthenticationRequestFingerPrintExtractorService and set to @Component

4_AuthenticationServiceImpl updated with injecting AuthenticationRequestFingerprintExtractor and extractAuthenticationRequestFingerprint() method implementation removed from it.

All is well by now (hopefully)