55th commit: account verification added
(next update project boundary management)

Account verification is a major concern in this project which is done based on UsernameType that can be either email or mobile type. All the required classes and interfaces related to account verification phase are located inside a new sub-package called verification package under authentication package.

I had two options here, the first one was jwt verification token generation which is stateless and sms otp (one-time password) 6-digit code which was saved into a redis cache with 1 minute time to leave.A set of environment variable related to jwt token generation and otpCode generation and caching are placed in the .yml file which are:

 verificationToken:
 otpCode:

In addition, io.jsonwebtoken required dependencies added top the .pom file.

1_ VerificationPayload record which will be sent to the kafka topic. (kafka functionality will be added to the project)

2_ JwtVerificationTokenService which implements JwtVerificationTokenGenerator and is responsible to generate new token , validate and extract username from the received token.

3_ JwtVerificationTokenProperties which is responsible for binding external configuration properties to pojo (plain old java object).

4_ otp package which contains Sms otp related classes and interfaces which is SOLID

a) Default6DigitSmsOtpSecurityCode is responsible to generate 6-digit random otpCode which implements SmsOtpSecurityCodeGenerator interface.

b) RedisCacheSmsOtpStore which is responsible to store otpCode in redis cache along with Time to leave and a related key which implements SmsOtpStore interface.

c) RedisCacheSmsOtpAttemptTracker is responsible to rateLimit (increments attempts ,block,reset cache) otpCode based on a specified number of attempts which implements SmsOtpAttemptTracker interface.

d)_ SmsOtpVerificationService which acts as an orchestrator service and is dependent on a number of components including SmsOtpSecurityCodeGenerator,SmsOtpStore,SmsOtpAttemptTracker as well as SmsOtpVerificationProperties. It has two method including generate and verify, the generate method calls generate method of smsOtpCodeGenerator and stores the otpCode into the redis cache by calling store method of the smsOtpStore. The verify method initially checks whether if the retrievedCode from the user who tries to verify his/her account is blocked or not, if it is not blocked it compare the retrieved key and the stored key in the cache and if they are both equal, it then removed the otpCode from the cache and reset the attempts, finally, if the otpCodes were not equal it increment the attempt by 1.

note: for security reasons,  otpCode and jwt verification token must be hashed.

5_ I added another strategy pattern which has two implementations, the first one is EmailVerificationStrategy and the second one is MobileVerificationStrategy , both implement VerificationStrategy, and they generate either otpCode or jwt token based one usernameType. The orchestrator class or the StrategyResolver here is called VerificationStrategyResolver class which has a list of VerificationStrategy and in its generatePayLoad() method it takes a username and UsernameType as its arguments, it converts the strategies to a stream and filters the first strategy that supports the usernameType and call the generate method of the strategy to create a VerificationPayload and if the strategy is empty it will throw an IllegalArgumentException.

6) verificationStrategyResolver will be called inside the register method of the AuthenticationServiceImpl class to generate a PayLoad and then publish an event and finally a VerificationEventPublisher based on kafka is implemented which needs to be taken care of.

7) OtpBlockedException is a custom exception which is handled in the CentralExceptionHandler class.