56th commit: A major update to manage package boundaries
As the project grew and became mature I decided to have a more clean boundaries between different packages and components, tried to stick to encapsulation standards and hiding classes, set most of the implementation classes package-private , renamed some components to be more readable, yet there are duplicate codes that are needed to be taken care of. Currently , I feel this package structure is more ideal than  the previous structure.